### 2025-07-28 #8657 @editorialbot
[PRE REVIEW]: batchplanet: Batch access and processing of PlanetScope imagery for spatiotemporal analysis in R
**Submitting author:** @yiluansong (<a href="http://orcid.org/0000-0003-3660-3797">Yiluan Song</a>)
**Repository:** https://github.com/zhulabgroup/batchplanet
**Branch with paper.md** (empty if default branch): release
**Version:** v1.0.0
