### 2025-08-29 bf71a34
Fix typo and add case for patial download
