+++ DESCRIPTION
@@ -13,8 +13,7 @@ Authors@R: c(
     person("Jiali", "Zhu", role = "aut",
            comment = c(ORCID = "0009-0009-7339-557X", affiliation = "School for Environment and Sustainability and Institute for Global Change Biology, University of Michigan, Ann Arbor, MI, USA")),
     person("Kai", "Zhu", role = "aut",
-           comment = c(ORCID = "0000-0003-1587-3317", affiliation = "School for Environment and Sustainability and Institute for Global Change Biology, University of Michigan, Ann Arbor, MI, USA"))    
-)
+           comment = c(ORCID = "0000-0003-1587-3317", affiliation = "School for Environment and Sustainability and Institute for Global Change Biology, University of Michigan, Ann Arbor, MI, USA")))
 Description: 
     The `batchplanet` R package provides a reproducible and scalable workflow for accessing and processing PlanetScope satellite imagery, enabling environmental researchers to efficiently perform spatiotemporal analysis of high-resolution remote sensing data. The package streamlines the steps required to work with the Planet API, including the ordering and downloading of imagery, retrieving and cleaning pixel-level time series, and computing derived metrics such as the Enhanced Vegetation Index (EVI) and green-up/down time. This package has supported peer-reviewed research in predicting reproductive phenology in wind-pollinated trees (Song et al., 2025). Generalizable beyond phenological research, this tool is particularly suited for environmental research that involves large volumes of PlanetScope imagery across multiple sites and long time periods.
 License: file LICENSE
