### 2025-09-03 ffc258d
Add guard
