### 2025-09-04 7f24d26
Add guard for NA min/max index
