### 2025-09-05 ff3212e
Add guard for NULL result
