### 2025-09-11 362a8ba
Update README with citation for BatchPlanet
Added citation for the original development of BatchPlanet.
