+++ README.md
@@ -1,6 +1,6 @@
 # BatchPlanet 🌿📊
 
-**BatchPlanet** is an R package designed to automate workflows for downloading, processing, and visualizing remote sensing imagery in batch from the Planet API. Originally developed to study tree phenology, the package has been generalized for broader environmental applications. It provides a suite of tools to:
+**BatchPlanet** is an R package designed to automate workflows for downloading, processing, and visualizing remote sensing imagery in batch from the Planet API. Originally developed to study tree phenology [Song et al., (2025)] (https://doi.org/10.1016/j.srs.2025.100205), the package has been generalized for broader environmental applications. It provides a suite of tools to:
 
 * **Download Data:** Retrieve PlanetScope images across multiple sites over time.
 * **Process Data:** Extract reflectance time series, calculate remote sensing indices (e.g., Enhanced Vegetation Index), and estimate transition time (e.g., green-up date).
