### 2025-10-07 99df3c4
case change, Fig2
