+++ tests/testthat/test-retrieve_planetscope_time_series.R
@@ -3,13 +3,13 @@
 # Unit tests for retrieve_planetscope_time_series_batch, retrieve_planetscope_time_series
 
 library(testthat)
-library(batchplanet)
+library(BatchPlanet)
 
 test_that("retrieve_planetscope_time_series_batch works with example data", {
   temp_dir <- withr::local_tempdir()
 
   # Test with example data
-  data_dir <- system.file("extdata/NEON/", package = "batchplanet")
+  data_dir <- system.file("extdata/NEON/", package = "BatchPlanet")
   dir.create(file.path(temp_dir, "raw", "SJER", "SJER_2025_60_90"), recursive = TRUE, showWarnings = FALSE)
   file.copy(
     from = list.files(file.path(data_dir, "raw", "SJER", "SJER_2025_60_90"), full.names = TRUE),
@@ -18,7 +18,7 @@ test_that("retrieve_planetscope_time_series_batch works with example data", {
   )
 
   # Run the function
-  df_coordinates <- read_csv(system.file("extdata", "NEON/example_neon_coordinates.csv", package = "batchplanet"), show_col_types = FALSE)
+  df_coordinates <- read_csv(system.file("extdata", "NEON/example_neon_coordinates.csv", package = "BatchPlanet"), show_col_types = FALSE)
 
   result <- retrieve_planetscope_time_series_batch(
     dir = temp_dir,
@@ -43,9 +43,9 @@ test_that("retrieve_planetscope_time_series_batch works with example data", {
 
 test_that("retrieve_planetscope_time_series works with example data", {
   # Test with example data
-  data_dir <- system.file("extdata/NEON/", package = "batchplanet")
+  data_dir <- system.file("extdata/NEON/", package = "BatchPlanet")
 
-  df_coordinates <- read_csv(system.file("extdata", "NEON/example_neon_coordinates.csv", package = "batchplanet"), show_col_types = FALSE)
+  df_coordinates <- read_csv(system.file("extdata", "NEON/example_neon_coordinates.csv", package = "BatchPlanet"), show_col_types = FALSE)
   df_coordinates_example <- df_coordinates %>% filter(site == "SJER", group == "Quercus")
 
   # Run the function
