+++ tests/testthat/test-read_data_product.R
@@ -3,10 +3,10 @@
 # Unit tests for read_data_product()
 
 library(testthat)
-library(batchplanet)
+library(BatchPlanet)
 
 test_that("read_data_product works with multiple product types", {
-  data_dir <- system.file("extdata/NEON/", package = "batchplanet")
+  data_dir <- system.file("extdata/NEON/", package = "BatchPlanet")
 
   df_ts <- read_data_product(data_dir, product_type = "ts")
   expect_s3_class(df_ts, "data.frame")
