+++ paper.md
@@ -1,228 +1,0 @@
----
-title: "*BatchPlanet*: Batch access and processing of PlanetScope imagery for spatiotemporal analysis in R"
-tags:
-  - R
-  - remote sensing
-  - PlanetScope
-  - phenology
-  - imagery processing
-  - time series
-authors:
-  - name: Yiluan Song
-    orcid: 0000-0003-3660-3797
-    affiliation: "1, 2"
-  - name: Alexandre R. Bevington
-    orcid: 0000-0003-1587-3317
-    affiliation: "3, 4"
-  - name: Ken Reid
-    orcid: 0000-0001-8654-2430
-    affiliation: 1
-  - name: Yi Liu
-    orcid: 0000-0002-5515-8804
-    affiliation: 2
-  - name: Jiali Zhu
-    orcid: 0009-0009-7339-557X
-    affiliation: 2
-  - name: Kai Zhu
-    orcid: 0000-0003-1587-3317
-    affiliation: 2
-affiliations:
-  - name: Michigan Institute for Data and AI in Society, University of Michigan, Ann Arbor, MI, USA
-    index: 1
-  - name: School for Environment and Sustainability and Institute for Global Change Biology, University of Michigan, Ann Arbor, MI, USA
-    index: 2
-  - name: Department of Geography, Earth and Environmental Sciences, University of Northern British Columbia, Prince George, BC, Canada
-    index: 3
-  - name: Ministry of Forests, Province of British Columbia, Prince George, BC, Canada
-    index: 4
-date: 2025-07-25
-repository: https://github.com/zhulabgroup/BatchPlanet
-bibliography: paper.bib
----
-
-# Summary
-
-The `BatchPlanet` R package provides a reproducible and scalable workflow for accessing and processing PlanetScope satellite imagery, enabling environmental researchers to efficiently perform spatiotemporal analysis of high-resolution remote sensing data. The package streamlines the steps required to work with the PlanetScope API, including the ordering and downloading of imagery, retrieving and cleaning pixel-level time series, and computing derived metrics such as the Enhanced Vegetation Index (EVI) and green-up/down time. This package has supported peer-reviewed research in predicting reproductive phenology in wind-pollinated trees [@song2025phenology]. Generalizable beyond phenological research, this tool is particularly suited for environmental research that involves large volumes of imagery across multiple sites and long time periods.
-
-# Statement of Need
-
-PlanetScope is a commercial high-resolution (~3-meter), near-daily satellite imagery product provided by Planet Labs. With its global coverage, high spatial resolution, and rapid revisit time, PlanetScope data has become increasingly valuable for scientific research and operational monitoring — enabling detailed analyses of land use change, ecosystem dynamics, disaster impacts, and more [@moon2021phenology]. While access to Planet data has been facilitated through the Planet API [@planet2017api], using such data still poses challenges: complex API interactions, large-volume data downloads, and non-trivial processing workflows often limit accessibility and reproducibility.
-
[…]
