+++ paper.bib
@@ -1,23 +1,0 @@
-@article{song2025phenology,
-  title={Predicting reproductive phenology of wind-pollinated trees via PlanetScope time series},
-  author={Song, Y. and Katz, D. S. and Zhu, Z. and Beaulieu, C. and Zhu, K.},
-  journal={Science of Remote Sensing},
-  year={2025},
-  pages={100205},
-  doi={10.1016/j.srs.2025.100205}
-}
-@article{moon2021phenology,
-  title={Multiscale assessment of land surface phenology from harmonized Landsat 8 and Sentinel-2, PlanetScope, and PhenoCam imagery},
-  author={Moon, M. and Richardson, A. D. and Friedl, M. A.},
-  journal={Remote Sensing of Environment},
-  volume={266},
-  pages={112716},
-  year={2021},
-  doi={10.1016/j.rse.2021.112716}
-}
-@misc{planet2017api,
-  author = {Planet Team},
-  title = {Planet application program interface: In space for life on Earth},
-  year = {2017},
-  howpublished = {\url{https://api.planet.com}}
-}
