+++ tests/testthat/setup.R
@@ -1,0 +1,2 @@
+message("Downloading sample data before running tests...")
+BatchPlanet::download_sample_data()
