+++ tests/testthat/test-visualize_time_series.R
@@ -7,14 +7,14 @@ library(BatchPlanet)
 
 test_that("visualize_time_series runs with NEON example data", {
   # Use a small subset of the NEON raw time series
-  df_ts <- system.file("extdata/NEON/ts/ts_SJER_Quercus.rds", package = "BatchPlanet") %>%
+  df_ts <- "sample-data/NEON/ts/ts_SJER_Quercus.rds" %>%
     read_rds() %>%
     sample_n(100)
   plt <- visualize_time_series(df_ts = df_ts, var = "green", ylab = "Green reflectance")
   expect_true("plotly" %in% class(plt))
 
   # Use a small subset of the NEON clean time series
-  df_clean <- system.file("extdata/NEON/clean/clean_SJER_Quercus.rds", package = "BatchPlanet") %>%
+  df_clean <- "sample-data/NEON/clean/clean_SJER_Quercus.rds" %>%
     read_rds() %>%
     sample_n(100)
   plt <- visualize_time_series(df_ts = df_clean, var = "evi", ylab = "EVI")
@@ -22,10 +22,10 @@ test_that("visualize_time_series runs with NEON example data", {
 
   # Use a small subset of the NEON clean time series with doy
   id_example <- c("NEON.PLA.D17.SJER.06001")
-  df_doy_sample <- system.file("extdata/NEON/doy/doy_SJER_Quercus.rds", package = "BatchPlanet") %>%
+  df_doy_sample <- "sample-data/NEON/doy/doy_SJER_Quercus.rds" %>%
     read_rds() %>%
     filter(id %in% id_example)
-  df_evi_sample <- system.file("extdata/NEON/clean/clean_SJER_Quercus.rds", package = "BatchPlanet") %>%
+  df_evi_sample <- "sample-data/NEON/clean/clean_SJER_Quercus.rds" %>%
     read_rds() %>%
     filter(id %in% id_example)
   plt <- visualize_time_series(df_ts = df_evi_sample, df_doy = df_doy_sample, var = "evi", ylab = "EVI", smooth = T)
