+++ tests/testthat/test-set_planetscope_parameters.R
@@ -18,11 +18,3 @@ test_that("set_api_key runs (skipped by default)", {
   # set_api_key()
   expect_true(TRUE)
 })
-
-test_that("set_data_directory runs (skipped by default)", {
-  skip_on_cran()
-  skip("User input test not run by default")
-  # Example (do not actually run):
-  # set_data_directory()
-  expect_true(TRUE)
-})
