+++ DESCRIPTION
@@ -43,4 +43,6 @@ Imports:
 Suggests:
     rmarkdown,
     knitr
-VignetteBuilder: knitr
+VignetteBuilder:
+    knitr,
+    rmarkdown
