+++ DESCRIPTION
@@ -21,14 +21,14 @@ Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
 Depends:
-    R (>= 3.3),
+    R (>= 3.3)
+Imports:
     tidyverse,
     sf,
     terra,
     foreach,
     doSNOW,
-    parallel
-Imports:
+    parallel,
     dotenv,
     ptw,
     imputeTS,
