+++ DESCRIPTION
@@ -21,14 +21,14 @@ Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
 Depends:
-    R (>= 3.3)
-Imports:
+    R (>= 3.3),
     tidyverse,
     sf,
     terra,
     foreach,
     doSNOW,
-    parallel,
+    parallel
+Imports:
     dotenv,
     ptw,
     imputeTS,
@@ -41,8 +41,8 @@ Imports:
     shiny,
     shinyWidgets
 Suggests:
-    rmarkdown,
-    knitr
+    knitr,
+    rmarkdown
 VignetteBuilder:
     knitr,
     rmarkdown
