### 2025-10-08 fc0100b
Add testthat to DESCRIPTION suggests
