### 2025-10-08 06dc110
Update VignetteBuilder in DESCRIPTION file
