+++ DESCRIPTION
@@ -45,6 +45,4 @@ Suggests:
     knitr,
     rmarkdown,
     testthat
-VignetteBuilder:
-    knitr,
-    rmarkdown
+VignetteBuilder: knitr
