### 2025-10-08 9b80b29
shorten paper
