### 2025-10-09 e43a6e0
static plot
