+++ tests/testthat/test-visualize_true_color_imagery.R
@@ -23,5 +23,5 @@ test_that("visualize_true_color_imagery runs with NEON example raster", {
 test_that("visualize_coordinates runs with NEON example data", {
   df_coordinates <- system.file("extdata/NEON/example_neon_coordinates.csv", package = "BatchPlanet") %>% readr::read_csv()
   plt <- visualize_coordinates(df_coordinates)
-  expect_true("plotly" %in% class(plt))
+  expect_true("plotly" %in% class(plt) || "gg" %in% class(plt) || "ggplot" %in% class(plt))
 })
