+++ vignettes/workflow.Rmd
@@ -116,7 +116,7 @@ At this point, you might have downloaded a large amount of images. Consider arch
 
 The `visualize_true_color_imagery_batch()` function starts a shiny app to visualize multiple images in the data directory. In the app, you can then choose the site and time you want to visualize and toggle brightness. This app might take a while to start.
 
-```{r}
+```{r, eval = F}
 visualize_true_color_imagery_batch(
   dir = dir_data_urban,
   # df_coordinates = df_coordinates_urban,
