+++ vignettes/workflow.Rmd
@@ -31,7 +31,7 @@ library(magrittr)
 In order to download PlanetScope imagery, you need to provide coordinates of points of interest. For this example, we will use several street tree inventories retrieved from the [OpenTrees.org](https://opentrees.org/).
 ```{r}
 # Read the data file with coordinates
-df_coordinates_urban <- readr::read_csv(system.file("extdata", "urban/example_urban_coordinates.csv", package = "BatchPlanet"), show_col_types = FALSE)
+df_coordinates_urban <- readr::read_csv(system.file("extdata", "urban", "example_urban_coordinates.csv", package = "BatchPlanet"), show_col_types = FALSE)
 head(df_coordinates_urban)
 ```
 You should create your own data frame with coordinates. Make sure your data frame for coordinates have the columns: `id` (unique id for each point of interest), `lon` for longitude, and `lat` for latitude. You may have a character column `site` if you have clustered coordinates at dispersed sites. You may also a character column `group` you can use to group coordinates later on.
@@ -41,6 +41,7 @@ What do we mean by "clustered coordinates at dispersed sites?" Use the function 
 visualize_coordinates(df_coordinates_urban %>% dplyr::sample_n(2000)) # Multiple sites across continental US
 visualize_coordinates(df_coordinates_urban %>% dplyr::filter(site == "NY") %>% dplyr::sample_n(2000)) # Zoom in to one site NY
 ```
+
 Why is it important to label the coordinates with sites? If we treat all coordinates as a single set and try to download PlanetScope imagery that cover them, we will end up with a tremendous amount of imagery that cover most of continental US. Instead, we will use the site labels to group the coordinates and create bounding boxes that cover each site. This way, we can download imagery that only cover the sites of interest. If your coordinates are already clustered in a small area, you don't have to supply the site labels. The package will process all coordinates as a single site.
 
 ## 2 Download PS data
@@ -73,14 +74,15 @@ setting <- set_planetscope_parameters(
 )
 ```
 
-Use the `set_data_directory()` function to specify a directory to store the downloaded data and later store processed data. It is recommended to use symbolic link to a directory on your HPC system. This way, you can easily access the data from your local machine without being limited by the storage space on your local machine. Please still closely monitor the storage space on your local machine or HPC system as the data can be large. For example, all images for New York City since 2017 can be over 1TB.
+Use the `download_sample_data()` function to download the sample data we will use in this vignette. For demonstration purpose, we are going to set the data directory to a subfolder of the `sample-data` folder.
+
+You should specify your own directory to store the downloaded data and later store processed data. It is recommended to use symbolic link to a directory on your HPC system. This way, you can easily access the data from your local machine without being limited by the storage space on your local machine. Please still closely monitor the storage space on your local machine or HPC system as the data can be large. For example, all images for New York City since 2017 can be over 1TB.
+
 ```{r}
 download_sample_data()
 dir_data <- "sample-data"
 dir_data_urban <- file.path(dir_data, "urban")
 ```
-
-For demonstration purpose, we are going to set the data directory to the package's extdata folder, but you should use `set_data_directory()` as above when you carry out your own analysis.
 
 ### 2.2 Order
 
@@ -114,7 +116,7 @@ At this point, you might have downloaded a large amount of images. Consider arch
 
 ### 2.4 Visualize true color imagery
 
[…]
