+++ README.md
@@ -58,7 +58,7 @@ To cite this package in publications use:
   Song Y, Bevington A, Reid K, Liu Y, Zhu J, Zhu K (2025).
   _BatchPlanet: Streamlined Processing and Analysis of PlanetScope and
   Satellite Imagery Data_. R package version 0.1.0,
-  <https://github.com/yiluansong/batchplanet>.
+  <https://github.com/zhulabgroup/BatchPlanet>.
 
 A BibTeX entry for LaTeX users is
 ```bibtex
@@ -68,7 +68,7 @@ Imagery Data},
     author = {Yiluan Song and Alexandre R. Bevington and Ken Reid and Yi Liu and Jiali Zhu and Kai Zhu},
     year = {2025},
     note = {R package version 0.1.0},
-    url = {https://github.com/yiluansong/batchplanet},
+    url = {https://github.com/zhulabgroup/BatchPlanet},
   }
 ```
 
