+++ README.md
@@ -56,15 +56,13 @@ devtools::test()
 To cite this package in publications use:
 
   Song Y, Bevington A, Reid K, Liu Y, Zhu J, Zhu K (2025).
-  _BatchPlanet: Streamlined Processing and Analysis of PlanetScope and
-  Satellite Imagery Data_. R package version 0.1.0,
+  _BatchPlanet: Batch access and processing of PlanetScope imagery for spatiotemporal analysis in R_. R package version 0.1.0,
   <https://github.com/zhulabgroup/BatchPlanet>.
 
 A BibTeX entry for LaTeX users is
 ```bibtex
   @Manual{,
-    title = {BatchPlanet: Streamlined Processing and Analysis of PlanetScope and Satellite
-Imagery Data},
+    title = {BatchPlanet: Batch access and processing of PlanetScope imagery for spatiotemporal analysis in R},
     author = {Yiluan Song and Alexandre R. Bevington and Ken Reid and Yi Liu and Jiali Zhu and Kai Zhu},
     year = {2025},
     note = {R package version 0.1.0},
