### 2025-10-10 3173667
Fix capitalization in DESCRIPTION for BatchPlanet
