+++ README.md
@@ -55,7 +55,7 @@ devtools::test()
 
 To cite this package in publications use:
 
-  Song Y, Bevington A, Reid K, Liu Y, Zhu J, Zhu K (2025).
+  Song Y, Bevington A, Reid K, Zhu J, Liu Y, Zhu K (2025).
   _BatchPlanet: Batch access and processing of PlanetScope imagery for spatiotemporal analysis in R_. R package version 0.1.0,
   <https://github.com/zhulabgroup/BatchPlanet>.
 
@@ -63,7 +63,7 @@ A BibTeX entry for LaTeX users is
 ```bibtex
   @Manual{,
     title = {BatchPlanet: Batch access and processing of PlanetScope imagery for spatiotemporal analysis in R},
-    author = {Yiluan Song and Alexandre R. Bevington and Ken Reid and Yi Liu and Jiali Zhu and Kai Zhu},
+    author = {Yiluan Song and Alexandre R. Bevington and Ken Reid and Jiali Zhu and Yi Liu and Kai Zhu},
     year = {2025},
     note = {R package version 0.1.0},
     url = {https://github.com/zhulabgroup/BatchPlanet},
