### 2025-12-17 #9268 @willgearty
Note, as per the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html), reviewers are expected to install the software they are reviewing and to verify the core functionality of the software. If this is not possible for the reviewers, we will unfortunately have to reject this submission, so I encourage @yiluansong to work with the reviewers to make sure they have API keys for Planet.
