### 2025-12-30 #9268 @QuinnAsena
The package seems to meet all the JOSS requirements (with the possible exception of Community Guidelines as noted below). The package is well structured and I did not encounter any bugs while testing the core functions. `batchplanet` has the potential to be very useful, filling the niche of R users accessing data from planet (as described in the accompanying paper), and can improve reproducibility for users. I'm sure `batchplanet` will grow with the user community and I think there is sufficient documentation for an initial release.

Suggested minor improvements:

- 'Community guidelines' could be stated in the README to satisfy JOSS requirements around how to contribute, report issues, and seek support.
- The three vignettes could each be fully self-contained. I found myself using some code from the 'workflow' vignette to fill in the other two vignettes. This could be a point of confusion for R users and I have raised an issue on GitHub to specify.
- The vignettes could be more visible in the README (e.g., include the code for launching each vignette [here](https://github.com/zhulabgroup/BatchPlanet?tab=readme-ov-file#example-workflow). Vignettes and tutorials will be important to users and should be maximally visible to help with user uptake.

The package is in a good state for a first release and JOSS publication.
