### 2025-12-30 37879b5
self-contained vignette
