### 2026-02-08 272a4f3
pipe operator
