+++ R/calculate_phenological_metrics.R
@@ -37,7 +37,6 @@
 #' )
 #' }
 #'
-#' @importFrom magrittr %>%
 #' @importFrom parallel makeCluster stopCluster
 #' @importFrom doSNOW registerDoSNOW
 #' @importFrom foreach foreach %dopar%
@@ -61,7 +60,7 @@ calculate_phenological_metrics_batch <- function(dir,
   # Create output directory for processed DOY data
   dir.create(file.path(dir, "doy"), showWarnings = FALSE)
 
-  v_file <- list.files(file.path(dir, "clean"), recursive = FALSE, full.names = FALSE) %>%
+  v_file <- list.files(file.path(dir, "clean"), recursive = FALSE, full.names = FALSE) |>
     filter_file_names(v_site, v_group)
 
   cl <- makeCluster(num_cores, outfile = "")
@@ -76,7 +75,7 @@ calculate_phenological_metrics_batch <- function(dir,
 
     df_doy <- calculate_phenological_metrics_sitegroup(df_index, df_thres, v_year, min_days, check_seasonality, var_index, extend_to_previous_year, extend_to_next_year)
 
-    f_doy <- file.path(dir, "doy", file %>% stringr::str_replace("clean_", "doy_"))
+    f_doy <- file.path(dir, "doy", file |> stringr::str_replace("clean_", "doy_"))
     readr::write_rds(df_doy, f_doy, compress = "gz")
   }
   stopCluster(cl)
@@ -84,35 +83,34 @@ calculate_phenological_metrics_batch <- function(dir,
   invisible(NULL)
 }
 
-#' @importFrom magrittr %>%
 calculate_phenological_metrics_sitegroup <- function(df_index, df_thres, v_year = NULL, min_days, check_seasonality, var_index = "evi", extend_to_previous_year = 275, extend_to_next_year = 90) {
   if (is.null(v_year)) {
-    v_year <- df_index %>%
-      dplyr::pull(year) %>%
-      unique() %>%
+    v_year <- df_index |>
+      dplyr::pull(year) |>
+      unique() |>
       sort()
   }
 
   ls_df_doy_year <- list()
   for (yearoi in v_year) {
-    df_index_year <- df_index %>%
-      dplyr::filter(year == yearoi | year == (yearoi - 1) | year == (yearoi + 1)) %>%
-      dplyr::filter(doy != 366) %>%
-      dplyr::mutate(doy = ifelse(doy >= extend_to_previous_year & year == yearoi - 1, doy - 365, doy)) %>%
-      dplyr::mutate(year = ifelse(doy <= 0 & year == yearoi - 1, year + 1, year)) %>%
-      dplyr::mutate(doy = ifelse(doy <= extend_to_next_year & year == yearoi + 1, doy + 365, doy)) %>%
-      dplyr::mutate(year = ifelse(doy > 365 & year == yearoi + 1, year - 1, year)) %>%
+    df_index_year <- df_index |>
+      dplyr::filter(year == yearoi | year == (yearoi - 1) | year == (yearoi + 1)) |>
+      dplyr::filter(doy != 366) |>
+      dplyr::mutate(doy = ifelse(doy >= extend_to_previous_year & year == yearoi - 1, doy - 365, doy)) |>
+      dplyr::mutate(year = ifelse(doy <= 0 & year == yearoi - 1, year + 1, year)) |>
+      dplyr::mutate(doy = ifelse(doy <= extend_to_next_year & year == yearoi + 1, doy + 365, doy)) |>
+      dplyr::mutate(year = ifelse(doy > 365 & year == yearoi + 1, year - 1, year)) |>
       dplyr::filter(year == yearoi)
 
-    v_id <- df_index %>%
[…]
