+++ DESCRIPTION
@@ -24,10 +24,9 @@ Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.2
 Depends:
-    R (>= 3.3),
+    R (>= 4.1),
 Imports:
     ggplot2,
-    magrittr,
     tidyr,
     dplyr,
     stringr,

[…]
