+++ NAMESPACE
@@ -29,6 +29,5 @@ import(shinyWidgets)
 importFrom(doSNOW,registerDoSNOW)
 importFrom(foreach,"%dopar%")
 importFrom(foreach,foreach)
-importFrom(magrittr,"%>%")
 importFrom(parallel,makeCluster)
 importFrom(parallel,stopCluster)
