### 2026-02-08 c958ddc
handling and documenting downloading errors
