### 2026-02-08 f8c03d0
visualize coordinates in leaflet
