### 2026-02-08 d1c6dcc
add NDVI
