+++ vignettes/workflow.Rmd
@@ -181,17 +181,20 @@ The `clean_planetscope_time_series_batch()` function removes low quality data in
 - The pixel was clear, had no snow, ice, shadow, haze, or cloud.
 - The usable data mask had algorithmic confidence in classification ≥ 80% for the pixel.
 
-The `calculate_evi` option allows you to calculate the Enhanced Vegetation Index (EVI) for the time series data and filter for EVI values between 0 and 1. The equation for EVI is as follows:
+The `calculate_index` option allows you to calculate the Normalized Difference Vegetation Index (NDVI) and/or Enhanced Vegetation Index (EVI) for the time series data. The equations for these two indices are as follows:
+$$
+NDVI = \frac{NIR - Red}{NIR + Red}
+$$
 $$
 EVI = 2.5 \times \frac{NIR - Red}{NIR + 6 \times Red - 7.5 \times Blue + 1}
 $$
-The function will create a folder `clean/` in the data directory you specified and save cleaned time series data with files names in the format of `clean_<site>_<group>.rds`.
+Users may manually compute additional indices using the returned surface reflectance bands (red, green, blue, nir). The function will create a folder `clean/` in the data directory you specified and save cleaned time series data with files names in the format of `clean_<site>_<group>.rds`.
 
 ```{r, eval = F}
-clean_planetscope_time_series_batch(dir = dir_data_NEON, v_site = c("HARV", "SJER"), v_group = c("Acer", "Quercus"), num_cores = 3, calculate_evi = T)
+clean_planetscope_time_series_batch(dir = dir_data_NEON, v_site = c("HARV", "SJER"), v_group = c("Acer", "Quercus"), num_cores = 3, calculate_index = "evi", filter_range = list(evi = c(0, 1)))
 ```
 
-Similar to Section 3.1, you can read in cleaned time series data using the `read_data_product()` function, with `product_type = "clean"`. You can then visualize the cleaned time series data using the `visualize_time_series()` function. The cleaned time series data will have a new column `evi` that we can visualize, if you set `calculate_evi = T` in the previous step.
+Similar to Section 3.1, you can read in cleaned time series data using the `read_data_product()` function, with `product_type = "clean"`. You can then visualize the cleaned time series data using the `visualize_time_series()` function. The cleaned time series data will have a new column `evi` that we can visualize, if you set `calculate_index = "evi"` in the previous step.
 
 ```{r}
 df_clean <- read_data_product(dir = dir_data_NEON, v_site = c("HARV", "SJER"), v_group = c("Quercus"), product_type = "clean")
