### 2026-02-08 88e49ef
generalize season metrics
