+++ R/calculate_season_metrics.R
@@ -1,7 +1,7 @@
-#' Calculate phenological metrics for time series from multiple sites and groups
+#' Calculate start/end of season metrics for time series from multiple sites and groups
 #'
 #' Reads remote sensing index time series files from the `clean/` subdirectory within the specified `dir`.
-#' For each site, group, and year combination, the function extends the time series to include the end of the previous year and the beginning of the next year (to capture early- and late-year phenological events), then calls `calculate_phenological_metrics()` to calculate phenological metrics for each time series.
+#' For each site, group, and year combination, the function extends the time series to include the end of the previous year and the beginning of the next year (to capture early- and late-year events), then calls `calculate_season_metrics()` to calculate start/end of season metrics for each time series, which are measured by the day-of-year (DOY) when the index first goes above or goes below specified threshold(s).
 #' Results are saved as `.rds` files under the `doy/` subdirectory within `dir`, prefixed `doy_`.
 #'
 #' @param dir Character. Base directory containing remote sensing index files (expects `clean/` subdirectory).
@@ -11,18 +11,18 @@
 #' @param df_thres Data frame of thresholds as from [set_thresholds()]; if `NULL`, uses default thresholds.
 #' @param var_index Character. Name of the remote sensing index column in the input data frame `df_index` to analyze (default: "evi").
 #' @param min_days Numeric. Minimum required number of non-NA data points in one year (default: 80).
-#' @param check_seasonality Logical. If `TRUE`, tests for significant seasonal changes before calculating phenological metrics (default: `TRUE`).
+#' @param check_seasonality Logical. If `TRUE`, tests for significant seasonal changes before calculating start/end of season metrics (default: `TRUE`).
 #' @param extend_to_previous_year Integer. Day of year to extend backward to, to capture early-year events (default: 275).
 #' @param extend_to_next_year Integer. Day of year to extend forward to, to capture late-year events (default: 90).
 #' @param num_cores Integer. Number of parallel workers for processing (default: 3).
 #'
-#' @return Invisibly returns `NULL` and saves calculated phenological metrics as `.rds` files in the `doy/` subdirectory of `dir`.
+#' @return Invisibly returns `NULL` and saves calculated start/end of season metrics as `.rds` files in the `doy/` subdirectory of `dir`.
 #'
 #' @examples
 #' \dontrun{
-#' # Example: Calculate phenological metrics for two sites and two groups
+#' # Example: Calculate start of season metrics for two sites and two groups
 #' df_thres <- set_thresholds(thres_up = c(0.3, 0.5))
-#' calculate_phenological_metrics_batch(
+#' calculate_season_metrics_batch(
 #'   dir = "alldata/PSdata/",
 #'   v_site = c("Site1", "Site2"),
 #'   v_group = c("Group1", "Group2"),
@@ -41,7 +41,7 @@
 #' @importFrom doSNOW registerDoSNOW
 #' @importFrom foreach foreach %dopar%
 #' @export
-calculate_phenological_metrics_batch <- function(dir,
+calculate_season_metrics_batch <- function(dir,
                                                  v_site = NULL,
                                                  v_group = NULL,
                                                  v_year = NULL,
@@ -73,7 +73,7 @@ calculate_phenological_metrics_batch <- function(dir,
     f_index <- file.path(dir, "clean", file)
     df_index <- readr::read_rds(f_index)
 
[…]
