+++ vignettes/customization.Rmd
@@ -117,12 +117,12 @@ df_ts_example <- read_data_product(dir = dir_data_NEON, v_site = "SJER", v_group
 df_clean_example <- clean_planetscope_time_series(df_ts = df_ts_example, calculate_index = "evi", filter_range = list(evi = c(0, 1)))
 ```
 
-## Calculate phenological metrics from a single data frame
+## Calculate start/end of season metrics from a single data frame
 
-You can customize your own phenological metrics calculation, not necessarily PlanetScope data or even EVI data, using the `calculate_phenological_metrics()` function that takes a data frame of time series data and a data frame of thresholds and returns a data frame with the calculated phenological metrics.
+You can customize your own start/end of season metrics calculation, not necessarily PlanetScope data or even EVI data, using the `calculate_season_metrics()` function that takes a data frame of time series data and a data frame of thresholds and returns a data frame with the calculated start/end of season metrics.
 
 ```{r}
 df_clean_example <- read_data_product(dir = dir_data_NEON, v_site = "SJER", v_group = "Quercus", product_type = "clean")
 df_thres <- set_thresholds(thres_up = c(0.3, 0.4, 0.5), thres_down = NULL)
-df_doy_example <- calculate_phenological_metrics(df_index = df_clean_example, df_thres = df_thres, var_index = "evi", min_days = 20, check_seasonality = F)
+df_doy_example <- calculate_season_metrics(df_index = df_clean_example, df_thres = df_thres, var_index = "evi", min_days = 20, check_seasonality = F)
 ```
