+++ NAMESPACE
@@ -1,7 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
-export(calculate_phenological_metrics)
-export(calculate_phenological_metrics_batch)
+export(calculate_season_metrics)
+export(calculate_season_metrics_batch)
 export(clean_planetscope_time_series)
 export(clean_planetscope_time_series_batch)
 export(determine_seasonality)

[…]
