+++ .github/workflows/test-coverage.yaml
@@ -23,6 +23,8 @@ jobs:
           needs: coverage
 
       - name: Test coverage
+        env:
+          CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
         run: |
           covr::codecov()
         shell: Rscript {0}
