+++ R/visualize_true_color_imagery.R
@@ -293,7 +293,7 @@ visualize_coordinates <- function(df_coordinates) {
   }
 
   df_coordinates <- df_coordinates |> tidyr::drop_na(lat, lon)
-  
+
   if (interactive()) {
     p <- leaflet::leaflet(df_coordinates) |>
       # leaflet::addTiles() |>   # Adds standard map background
