+++ _pkgdown.yml
@@ -1,0 +1,4 @@
+url: https://zhulabgroup.github.io/BatchPlanet/
+template:
+  bootstrap: 5
+
