+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: BatchPlanet
 Title: Batch access and processing of PlanetScope imagery for spatiotemporal analysis in R
-Version: 0.1.0
+Version: 0.2.0
 Date: 2025-10-08
 Authors@R: c(
     person("Yiluan", "Song", email = "songyl@umich.edu", role = c("aut", "cre"),
@@ -17,7 +17,7 @@ Authors@R: c(
            comment = c(ORCID = "0000-0003-1587-3317", affiliation = "School for Environment and Sustainability and Institute for Global Change Biology, University of Michigan, Ann Arbor, MI, USA")))
 Description: 
     The `BatchPlanet` R package provides a reproducible and scalable workflow for accessing and processing PlanetScope satellite imagery, enabling environmental researchers to efficiently perform spatiotemporal analysis of high-resolution remote sensing data. The package streamlines the steps required to work with the Planet API, including the ordering and downloading of imagery, retrieving and cleaning pixel-level time series, and computing derived metrics such as the Enhanced Vegetation Index (EVI) and green-up/down time. This package has supported peer-reviewed research in predicting reproductive phenology in wind-pollinated trees (Song et al., 2025). Generalizable beyond phenological research, this tool is particularly suited for environmental research that involves large volumes of PlanetScope imagery across multiple sites and long time periods.
-URL: https://yiluansong.r-universe.dev/BatchPlanet
+URL: https://yiluansong.r-universe.dev/BatchPlanet, https://zhulabgroup.github.io/BatchPlanet/
 BugReports: https://github.com/zhulabgroup/BatchPlanet/issues
 License: file LICENSE
 Encoding: UTF-8
