+++ paper.md
@@ -1,0 +1,291 @@
+# Summary
+
+The `BatchPlanet` R package offers a reproducible and scalable workflow
+for accessing and processing PlanetScope satellite imagery. It is
+designed for environmental researchers to efficiently perform
+spatiotemporal analysis on high-resolution remote sensing data. The
+package streamlines imagery ordering and downloading with the
+PlanetScope API, retrieval and cleaning of pixel-level time series,
+calculation of vegetation indices like the Enhanced Vegetation Index
+(EVI), and calculation of start/end of season metrics. This tool is
+particularly valuable for research involving large volumes of imagery
+across multiple sites and extended time periods.
+
+# Statement of Need
+
+PlanetScope imagery provides global, high-resolution (~3-meter),
+near-daily data, making it valuable for scientific research and
+monitoring of phenology \[@moon2021phenology\], land use change,
+disaster impacts, and more. While the Planet API facilitates access
+\[@planet2017api\], using this data remains challenging due to complex
+API interactions, limits on large-volume data downloads, and non-trivial
+processing workflows, which can hinder reproducibility.
+
+Existing options include the official [Planet Python
+SDK](https://planet-sdk-for-python-v2.readthedocs.io/en/latest/python/sdk-guide/),
+cloud-based platforms like [Sentinel
+Hub](https://www.sentinel-hub.com/develop/api/) and [Google Earth Engine
+(GEE)](https://docs.planet.com/platform/integrations/google-earth-engine/),
+and the R package `planetR` \[@bevington2024planetr\]. However, the
+Python and JavaScript used by the first three platforms may be less
+familiar to R users. Cloud platforms also restrict user control over
+processing environments and local data storage. Furthermore, existing
+tools typically require users to write extensive custom scripts for
+batch downloading and processing across multiple sites, a process often
+limited by computing and storage resources. `BatchPlanet` addresses
+these gaps by providing an R-native tool for batch access and processing
+of PlanetScope imagery (Table 1). Its streamlined, parallelized
+functions are designed for scalability, transparency, and
+reproducibility in scientific data pipelines. This package has supported
+peer-reviewed research in predicting reproductive phenology in
+wind-pollinated trees \[@song2025phenology\].
+
+| Feature / Tool                     | **Planet Python SDK** | **Sentinel Hub**               | **Google Earth Engine (GEE)** | **planetR (Bevington)** | **BatchPlanet** |
+|------------------------------------|-----------------------|--------------------------------|-------------------------------|-------------------------|-----------------|
+| **Primary Language**               | Python                | Python                         | JavaScript / Python           | R                       | R               |
+| **Processing Environment**         | Local/Cloud           | Cloud                          | Cloud                         | Local                   | Local           |
+| **Data Control & Reproducibility** | High                  | Moderate                       | Low                           | High                    | High            |
+| **Batch Processing**               | Via scripting/CLI     | Supported for enterprise users | Via scripting                 | Via scripting           | Streamlined     |
+| **Time Series Analysis Tools**     | Not supported         | Limited                        | Supported                     | Not supported           | Supported       |
[…]
