+++ .github/workflows/test-coverage.yaml
@@ -1,4 +1,4 @@
-name: test-coverage
+name: Test coverage
 
 on: [push]
 
