+++ vignettes/1_workflow.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "1 Workflow"
+title: "1 Getting Started Workflow"
 author: "Yiluan Song"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
