+++ _pkgdown.yml
@@ -1,4 +1,3 @@
 url: https://zhulabgroup.github.io/BatchPlanet/
 template:
   bootstrap: 5
-
