+++ .github/workflows/pkgdown.yaml
@@ -10,8 +10,6 @@ on:
     types: [published]
   workflow_dispatch:
 
-name: pkgdown.yaml
-
 permissions: read-all
 
 jobs:
