+++ paper.md
@@ -92,7 +92,7 @@ install.packages('BatchPlanet',
 library(BatchPlanet)
 ```
 
-To facilitate adoption and ensure reproducibility, `BatchPlanet` provides three vignettes that demonstrate its full feature set: [**Getting started workflow**](https://zhulabgroup.github.io/BatchPlanet/articles/1_workflow.html), [**Customization and advanced usage**](https://zhulabgroup.github.io/BatchPlanet/articles/2_customization.html), and [**Miscellaneous time series processing tools**](https://zhulabgroup.github.io/BatchPlanet/articles/3_tools.html). We highlight some key functions below.
+To facilitate adoption and ensure reproducibility, `BatchPlanet` provides three vignettes that demonstrate its full feature set: [**Getting started workflow**](https://zhulabgroup.github.io/BatchPlanet/articles/workflow.html), [**Customization and advanced usage**](https://zhulabgroup.github.io/BatchPlanet/articles/customization.html), and [**Miscellaneous time series processing tools**](https://zhulabgroup.github.io/BatchPlanet/articles/tools.html). We highlight some key functions below.
 
 Read example coordinates.
 
@@ -123,7 +123,7 @@ dir_data_NEON <- file.path(dir_data, "NEON")
 
 Order and download imagery.
 
-Note: Before proceeding to downloading, users should inspect their [Planet account](https://www.planet.com/account/) to confirm that all orders reached a "success" status. Failed orders will result in errors during downloading. Refer to the [package vignette](https://zhulabgroup.github.io/BatchPlanet/articles/1_workflow.html) for troubleshooting tips when orders fail.
+Note: Before proceeding to downloading, users should inspect their [Planet account](https://www.planet.com/account/) to confirm that all orders reached a "success" status. Failed orders will result in errors during downloading. Refer to the [package vignette](https://zhulabgroup.github.io/BatchPlanet/articles/workflow.html) for troubleshooting tips when orders fail.
 
 ``` r
 order_planetscope_imagery_batch(
