+++ paper.md
@@ -141,11 +141,11 @@ library(BatchPlanet)
 To facilitate adoption and ensure reproducibility, `BatchPlanet`
 provides three vignettes that demonstrate its full feature set:
 [**Getting started
-workflow**](https://zhulabgroup.github.io/BatchPlanet/articles/1_workflow.html),
+workflow**](https://zhulabgroup.github.io/BatchPlanet/articles/workflow.html),
 [**Customization and advanced
-usage**](https://zhulabgroup.github.io/BatchPlanet/articles/2_customization.html),
+usage**](https://zhulabgroup.github.io/BatchPlanet/articles/customization.html),
 and [**Miscellaneous time series processing
-tools**](https://zhulabgroup.github.io/BatchPlanet/articles/3_tools.html).
+tools**](https://zhulabgroup.github.io/BatchPlanet/articles/tools.html).
 We highlight some key functions below.
 
 Read example coordinates.
@@ -181,7 +181,7 @@ Note: Before proceeding to downloading, users should inspect their
 [Planet account](https://www.planet.com/account/) to confirm that all
 orders reached a “success” status. Failed orders will result in errors
 during downloading. Refer to the [package
-vignette](https://zhulabgroup.github.io/BatchPlanet/articles/1_workflow.html)
+vignette](https://zhulabgroup.github.io/BatchPlanet/articles/workflow.html)
 for troubleshooting tips when orders fail.
 
 ``` r
