+++ articles/index.md
@@ -2,9 +2,9 @@
 
 ### All vignettes
 
-- [1 Getting Started
-  Workflow](https://zhulabgroup.github.io/BatchPlanet/articles/1_workflow.md):
 - [2 Customization and advanced
-  usage](https://zhulabgroup.github.io/BatchPlanet/articles/2_customization.md):
+  usage](https://zhulabgroup.github.io/BatchPlanet/articles/customization.md):
 - [3 Miscellaneous time series processing
-  tools](https://zhulabgroup.github.io/BatchPlanet/articles/3_tools.md):
+  tools](https://zhulabgroup.github.io/BatchPlanet/articles/tools.md):
+- [1 Getting Started
+  Workflow](https://zhulabgroup.github.io/BatchPlanet/articles/workflow.md):
