+++ llms.txt
@@ -59,15 +59,15 @@ For complete guides on using all features of `BatchPlanet`, please
 explore our package vignettes:
 
 - [**Getting started
-  workflow**](https://zhulabgroup.github.io/BatchPlanet/articles/1_workflow.html):
+  workflow**](https://zhulabgroup.github.io/BatchPlanet/articles/workflow.html):
   A complete, step-by-step example demonstrating how to download
   PlanetScope data and process time series.
 - [**Customization and advanced
-  usage**](https://zhulabgroup.github.io/BatchPlanet/articles/2_customization.html):
+  usage**](https://zhulabgroup.github.io/BatchPlanet/articles/customization.html):
   Advanced configurations for tailoring the processing pipeline to
   specific research needs.
 - [**Miscellaneous time series processing
-  tools**](https://zhulabgroup.github.io/BatchPlanet/articles/3_tools.html):
+  tools**](https://zhulabgroup.github.io/BatchPlanet/articles/tools.html):
   Two generalizable tools for time series analysis.
 
 Full documentation of functions can be found on the [Reference
@@ -215,9 +215,9 @@ Enjoy exploring and analyzing remote sensing data with BatchPlanet! 🚀
 
 ### All vignettes
 
-- [1 Getting Started
-  Workflow](https://zhulabgroup.github.io/BatchPlanet/articles/1_workflow.md):
 - [2 Customization and advanced
-  usage](https://zhulabgroup.github.io/BatchPlanet/articles/2_customization.md):
+  usage](https://zhulabgroup.github.io/BatchPlanet/articles/customization.md):
 - [3 Miscellaneous time series processing
-  tools](https://zhulabgroup.github.io/BatchPlanet/articles/3_tools.md):
+  tools](https://zhulabgroup.github.io/BatchPlanet/articles/tools.md):
+- [1 Getting Started
+  Workflow](https://zhulabgroup.github.io/BatchPlanet/articles/workflow.md):

[…]
