### 2026-02-09 #9268 @yiluansong
Here are my responses to @QuinnAsena's comments and suggestions.
- I have added a section of 'Community guidelines' in the README [here](https://github.com/zhulabgroup/BatchPlanet/blob/1b1cb2e8d50d633fcd6292f9d772038c01b7a964/README.md?plain=1#L85-L97).
- The three vignettes have been revised to be fully self-contained in [this commit](https://github.com/zhulabgroup/BatchPlanet/commit/37879b5eea0febb4dd156931e1728e8abf9817a4). The vignettes are also rendered and published via pkgdown, confirming that they are self-contained.
- I have increased the visiblity of the three vignettes in three ways: 1) publishing a [pkgdown package website](https://zhulabgroup.github.io/BatchPlanet/index.html) with rendered vignettes, 2) referencing links to vignettes in the [README](https://github.com/zhulabgroup/BatchPlanet/blob/1b1cb2e8d50d633fcd6292f9d772038c01b7a964/README.md?plain=1#L41-L47), and 3) referencing links to vignettes in the [paper](https://github.com/zhulabgroup/BatchPlanet/blob/1b1cb2e8d50d633fcd6292f9d772038c01b7a964/paper.md?plain=1#L95).
