### 2026-02-09 #9268 @yiluansong
Here are my responses to @ajpelu's comments and suggestions.

- I have increased the visiblity of the three vignettes in three ways: 1) publishing a [pkgdown package website](https://zhulabgroup.github.io/BatchPlanet/index.html) with rendered vignettes, 2) referencing links to vignettes in the [README](https://github.com/zhulabgroup/BatchPlanet/blob/1b1cb2e8d50d633fcd6292f9d772038c01b7a964/README.md?plain=1#L41-L47), and 3) referencing links to vignettes in the [paper](https://github.com/zhulabgroup/BatchPlanet/blob/1b1cb2e8d50d633fcd6292f9d772038c01b7a964/paper.md?plain=1#L95).
- The new [pkgdown package website](https://zhulabgroup.github.io/BatchPlanet/index.html) also increases visibility and easy access of the documentation of functions.
- I identified the estimation of phenological metrics to be overly field-specific. I have pivoted this component to estimate start and/or end of season metrics, which is more generally applicable to diverse fields. I have made changes in the following ways: 1) renaming the functions, 2) removing variable names specific to phenology (e.g., "green-up/down"), 3) improving function documentation to use general language, and 4) rephrasing relevant statements in the vignettes and paper. Most changes were in [this commit](https://github.com/zhulabgroup/BatchPlanet/commit/88e49ef13a86e5e435441ef4cd19c95e666d2089).
- I have added the functionality to calculate vegetation indices NDVI in addition to EVI. The current functions allow the user to use an argument like `calculate_index = c("ndvi", "evi")` to specify the indices to calculate. I also added optional filters to filter out indices that are outside user-specified ranges. Most changes were in [this commit](https://github.com/zhulabgroup/BatchPlanet/commit/d1c6dcc111c5da298448a28e4bbbd9b1cac32c63).
- I have included a [badge](https://github.com/zhulabgroup/BatchPlanet/blob/1b1cb2e8d50d633fcd6292f9d772038c01b7a964/README.md?plain=1#L60) for test coverage from Codecov. The current test coverage is 55%, mainly because many functions could not be tested without a valid API key.
- I have taken the suggestion of using `leaflet` in the function `visualize_coordinates()` to provide more information. Using a basemap of ESRI world imagery, the user can now not only inspect the spatial clustering of coordinates, but also inspect how coordinates correspond to real-world objects. Most changes were in [this commit](https://github.com/zhulabgroup/BatchPlanet/commit/f8c03d070457577c22e13aebf695f773c9a231ea).
-  I have added more details in the function-level documentation for `set_planetscope_parameters()`.  See [this commit](https://github.com/zhulabgroup/BatchPlanet/commit/8cb4376634419b28bab568aeb3bd9212abc340b0).
- I have replaced all pipe operators with the native R pipe `|>` and reduced the dependency on `magrittr`. Because of this, the required R version is now 4.1+. Most changes were in [this commit](https://github.com/zhulabgroup/BatchPlanet/commit/272a4f3b52d4af9a574b2ff1968df4321097c8cb).
- I have added [text](https://github.com/zhulabgroup/BatchPlanet/blob/1b1cb2e8d50d633fcd6292f9d772038c01b7a964/paper.md?plain=1#L126) in the paper to explain how to inspect orders in Planet website to minimize errors in downloading, and to reference published vignettes for more troubleshooting tips. I have also substantially revised how errors in downloading are handed in [this commit](https://github.com/zhulabgroup/BatchPlanet/commit/c958ddc98ad47c667c3afcc8d40f7dc818cd3e0f), providing more detailed error messages and avoiding infinite loops.
- I was not able to reproduce the typographical issues in `read_data_product()`, but I applied the `styler` package to tidy up this file.
- In the paper, I have replaced “derived metrics” with “vegetation indices” to be more accurate.
- The previously used "green-up and green-down" metrics were replaced by "start/end of season" metrics. I have added text in the paper ([example](https://github.com/zhulabgroup/BatchPlanet/blob/1b1cb2e8d50d633fcd6292f9d772038c01b7a964/paper.md?plain=1#L76)) and function-level documentation ([example](https://github.com/zhulabgroup/BatchPlanet/blob/1b1cb2e8d50d633fcd6292f9d772038c01b7a964/R/calculate_season_metrics.R#L4)) to explain the rationale.
- The [caption of Fig. 2](https://github.com/zhulabgroup/BatchPlanet/blob/1b1cb2e8d50d633fcd6292f9d772038c01b7a964/paper.md?plain=1#L82) has been substantially revised to explain the green shades and lines.
- The incorrectly formatted list in the paper has been fixed.
- I have added citations for the previous R package for Planet data, the calculation of vegetation indices, the rationale of start/end of season metrics, and provided links to other existing Planet tools.
