### 2026-02-09 #9268 @willgearty
Thank you @yiluansong for addressing the two reviewers' comments and for your thorough responses. It looks like the software and paper have both improved through this process (and I am very happy to see the shiny new pkgdown website!).

@ajpelu @QuinnAsena please review the changes and update your review checklists accordingly. If your checklists are complete and your are satisfied with the state of the software and paper, please post a message to this effect. If not, please post a message with further comments/requests.
