### 2026-02-09 #9268 @willgearty
@yiluansong Please add the recommended DOIs as listed above in the meantime.
