### 2026-02-11 7fbe130
Update bibliography by removing and modifying entries
Removed the article entry for 'Multiscale assessment of land surface phenology' and added DOI for the 'Overview of the radiometric and biophysical performance of the MODIS vegetation indices' article.
