+++ paper.bib
@@ -4,15 +4,6 @@
   url = {https://github.com/bevingtona/planetR},
   year = {2019}
 }
-@article{moon2021multiscale,
-  title={Multiscale assessment of land surface phenology from harmonized Landsat 8 and Sentinel-2, PlanetScope, and PhenoCam imagery},
-  author={Moon, Minkyu and Richardson, Andrew D and Friedl, Mark A},
-  journal={Remote Sensing of Environment},
-  volume={266},
-  pages={112716},
-  year={2021},
-  publisher={Elsevier}
-}
 @article{huete2002overview,
   title={Overview of the radiometric and biophysical performance of the MODIS vegetation indices},
   author={Huete, Alfredo and Didan, Kamel and Miura, Tomoaki and Rodriguez, E Patricia and Gao, Xiang and Ferreira, Laerte G},
@@ -21,7 +12,8 @@
   number={1-2},
   pages={195--213},
   year={2002},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/S0034-4257(02)00096-2}
 }
 @article{song2025phenology,
   title={Predicting reproductive phenology of wind-pollinated trees via PlanetScope time series},
