+++ pkgdown.yml
@@ -5,7 +5,7 @@ articles:
   customization: customization.html
   tools: tools.html
   workflow: workflow.html
-last_built: 2026-02-11T03:32Z
+last_built: 2026-02-11T03:38Z
 urls:
   reference: https://zhulabgroup.github.io/BatchPlanet/reference
   article: https://zhulabgroup.github.io/BatchPlanet/articles
