### 2026-02-16 #9268 @ajpelu
Dear editor @willgearty, the authors have adequately addressed the comments and suggestions, and in my view the manuscript has improved substantially. The inclusion of vignettes and the website has made the R package much more attractive and usable.

However, I suggest revising a few minor presentation issues:

- Remove the duplicate caption titles for Table 1, Figure 1, and Figure 2.

- In Table 1, it would be helpful to separate the rows more clearly to improve readability.

- For Figure 1, please reduce the size to improve visualization when rendered as a PDF, as the caption is currently difficult to read.

- In Figure 2, although the legend indicates that the vertical lines correspond to 30%, 40%, and 50%, it is not clear which line corresponds to each value in the plots, which may be confusing.

From my side, I recommend accepting the manuscript once these minor issues have been addressed. Congratulations @yiluansong
