### 2026-02-26 #9268 @QuinnAsena
Dear @willgearty, the update to batchplanet, including a website and updated README much improve the visibility and usability of the package. Additional functionality has also been included. I would be very happy to see this published. Congrats @yiluansong and team, apologies for any delay on my part.

I agree with the [remaining comments](https://github.com/openjournals/joss-reviews/issues/9268#issuecomment-3907856260) from @ajpelu for a few minor adjustments to the manuscript.
