### 2026-03-05 cef294f
visualization debug
