### 2026-03-05 2f0a144
Adjust array stretch factor in LaTeX header
