### 2026-03-05 #9268 @yiluansong
Dear editor @willgearty  and reviewers @ajpelu @QuinnAsena, I have made the following minor revisions to the paper.

- The duplicate caption titles for Table 1, Figure 1, and Figure 2 were removed in the previous revision, but the revision did not show up last time for some reason. I have confirmed that the duplicated captions have been removed in the latest version.
- In Table 1, I increased the spacing between the rows to better separate them visually.
- I reduced the size of Figure 1 and improved its layout and resolution. I did not reduce its size even further due to typesetting reasons.
- In Figure 2, I added an explanation that the three vertical lines correspond to 30%, 40%, and 50% from left to right. I also debugged the function for visualization, improved the figure resolution, and added a demonstration of the interaction with ggplot.

Thank you!
