+++ DESCRIPTION
@@ -1,12 +1,12 @@
 Package: BatchPlanet
 Title: Batch access and processing of PlanetScope imagery for spatiotemporal analysis in R
-Version: 0.2.0
+Version: 1.0.0
 Date: 2025-10-08
 Authors@R: c(
     person("Yiluan", "Song", email = "songyl@umich.edu", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-3660-3797", affiliation = "Michigan Institute for Data and AI in Society, University of Michigan, Ann Arbor, MI, USA; School for Environment and Sustainability and Institute for Global Change Biology, University of Michigan, Ann Arbor, MI, USA")),
     person("Alexandre R.", "Bevington", role = "aut",
-           comment = c(ORCID = "0000-0003-1587-3317", affiliation = "Department of Geography, Earth and Environmental Sciences, University of Northern British Columbia, Prince George, BC, Canada; Ministry of Forests, Province of British Columbia, Prince George, BC, Canada")),
+           comment = c(ORCID = "0000-0002-9249-4444", affiliation = "Department of Geography, Earth and Environmental Sciences, University of Northern British Columbia, Prince George, BC, Canada; Ministry of Forests, Province of British Columbia, Prince George, BC, Canada")),
     person("Kenneth N.", "Reid", role = "aut",
            comment = c(ORCID = "0000-0001-8654-2430", affiliation = "Honorary Fellow - University of Stirling, Scotland, UK")),
     person("Jiali", "Zhu", role = "aut",
@@ -14,7 +14,7 @@ Authors@R: c(
     person("Yi", "Liu", role = "aut",
            comment = c(ORCID = "0000-0002-5515-8804", affiliation = "School for Environment and Sustainability and Institute for Global Change Biology, University of Michigan, Ann Arbor, MI, USA")),
     person("Kai", "Zhu", role = "aut",
-           comment = c(ORCID = "0000-0003-1587-3317", affiliation = "School for Environment and Sustainability and Institute for Global Change Biology, University of Michigan, Ann Arbor, MI, USA")))
+           comment = c(ORCID = "0000-0003-1587-3317", affiliation = "School for Environment and Sustainability and Institute for Global Change Biology, University of Michigan, Ann Arbor, MI, USA; Michigan Institute for Data and AI in Society, University of Michigan, Ann Arbor, MI, USA")))
 Description: 
     The `BatchPlanet` R package provides a reproducible and scalable workflow for accessing and processing PlanetScope satellite imagery, enabling environmental researchers to efficiently perform spatiotemporal analysis of high-resolution remote sensing data. The package streamlines the steps required to work with the Planet API, including the ordering and downloading of imagery, retrieving and cleaning pixel-level time series, and computing derived metrics such as the Enhanced Vegetation Index (EVI) and green-up/down time. This package has supported peer-reviewed research in predicting reproductive phenology in wind-pollinated trees (Song et al., 2025). Generalizable beyond phenological research, this tool is particularly suited for environmental research that involves large volumes of PlanetScope imagery across multiple sites and long time periods.
 URL: https://yiluansong.r-universe.dev/BatchPlanet, https://zhulabgroup.github.io/BatchPlanet/
