### 2026-03-10 37a41df
figure display
