+++ DESCRIPTION
@@ -6,7 +6,7 @@ Authors@R: c(
     person("Yiluan", "Song", email = "songyl@umich.edu", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0003-3660-3797", affiliation = "Michigan Institute for Data and AI in Society, University of Michigan, Ann Arbor, MI, USA; School for Environment and Sustainability and Institute for Global Change Biology, University of Michigan, Ann Arbor, MI, USA")),
     person("Alexandre R.", "Bevington", role = "aut",
-           comment = c(ORCID = "0000-0002-9249-4444", affiliation = "Department of Geography, Earth and Environmental Sciences, University of Northern British Columbia, Prince George, BC, Canada; Ministry of Forests, Province of British Columbia, Prince George, BC, Canada")),
+           comment = c(ORCID = "0000-0002-9249-4444", affiliation = "Department of Geography, Earth and Environmental Sciences, University of Northern British Columbia, Prince George, BC, Canada")),
     person("Kenneth N.", "Reid", role = "aut",
            comment = c(ORCID = "0000-0001-8654-2430", affiliation = "Honorary Fellow - University of Stirling, Scotland, UK")),
     person("Jiali", "Zhu", role = "aut",
