### 2026-03-11 #9268 @yiluansong
The latest release to be used for JOSS paper is v1.0.0.
Zenodo DOI is https://doi.org/10.5281/zenodo.18963802.
@willgearty I believe that I have completed all items on the author checklist. Thank you!
